+++ src/media/visual/image/sixel/palette.rs
@@ -7,7 +7,17 @@ use crate::{NotEnoughSpace, SixelColor};
 
 #[doc = crate::_tags!(color term)]
 /// Palette of Sixel colors with fixed capacity.
-#[doc = crate::_doc_meta!{location("media/visual/image")}]
+#[doc = crate::_doc_meta!{
+    location("media/visual/image", struct SixelPalette),
+    #[cfg(target_pointer_width = "32")]
+    test_size_of(__: SixelPalette<0> = 4|32; niche !Option),
+    #[cfg(target_pointer_width = "32")]
+    test_size_of(__: SixelPalette<255> = 1028|8192; niche Option),
+    #[cfg(target_pointer_width = "64")]
+    test_size_of(__: SixelPalette<0> = 8|64; niche !Option),
+    #[cfg(target_pointer_width = "64")]
+    test_size_of(__: SixelPalette<255> = 1032|8256; niche Option),
+}]
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
 pub struct SixelPalette<const CAP: usize> {
     colors: [Option<SixelColor>; CAP],
@@ -198,7 +208,7 @@ impl<'a, const CAP: usize> Iterator for SixelPaletteIter<'a, CAP> {
 }
 
 #[cfg(test)]
-mod tests {
+mod _test {
     #[test]
     fn write_definitions() {}
 }